chore: Sketch out simplified data source architecture
