--- site/dataSources/file.ts
@@ -1,1 +1,0 @@
-export default Deno.readTextFile;
