--- site/dataSources/index.ts
@@ -1,4 +1,0 @@
-import file from "./file.ts";
-import indexMarkdown from "./indexMarkdown.ts";
-
-export { file, indexMarkdown };
