+++ site/routes.json
@@ -1,26 +1,12 @@
 {
   "/": {
+    "layout": "siteIndex",
     "meta": {
       "title": "Gustwind",
       "description": "Deno powered JSON oriented site generator"
     },
-    "layout": "siteIndex",
-    "dataSources": [
-      {
-        "id": "readme",
-        "operation": "file",
-        "input": "./README.md",
-        "transformWith": ["markdown"]
-      }
-    ],
+    "dataSources": ["readme"],
     "expand": {
-      "dataSources": [
-        {
-          "id": "documentation",
-          "operation": "indexMarkdown",
-          "input": "./documentation"
-        }
-      ],
       "matchBy": {
         "dataSource": "documentation",
         "slug": "data.slug"
@@ -38,21 +24,8 @@
       "title": "Blog",
       "description": "A blog about Gustwind"
     },
-    "dataSources": [
-      {
-        "id": "blogPosts",
-        "operation": "indexMarkdown",
-        "input": "./blogPosts"
-      }
-    ],
+    "dataSources": ["blogPosts"],
     "expand": {
-      "dataSources": [
-        {
-          "id": "blogPosts",
-          "operation": "indexMarkdown",
-          "input": "./blogPosts"
-        }
-      ],
       "matchBy": {
         "dataSource": "blogPosts",
         "slug": "data.slug"
@@ -92,17 +65,11 @@
   },
   "atom.xml": {
     "layout": "rssPage",
-    "type": "xml",
     "meta": {
       "title": "Gustwind",
       "description": "Gustwind blog"
     },
-    "dataSources": [
-      {
-        "id": "blogPosts",
-        "operation": "indexMarkdown",
-        "input": "./blogPosts"
-      }
-    ]
+    "dataSources": ["blogPosts"],
+    "type": "xml"
   }
 }
